fix: convert lat and lon to float instead of int

+++ src/tile/Util.h
@@ -26,8 +26,8 @@ struct Coordinate {
 
 float deg2Rad(float degree);
 float rad2Deg(float radius);
-int longitude2X(float longitude, int zoom);
-int latitude2Y(float latitude, int zoom);
+float longitude2X(float longitude, int zoom);
+float latitude2Y(float latitude, int zoom);
 float x2Longitude(float x, int zoom);
 float y2Latitude(float y, int zoom);
 int bestZoomLevel(const BoundingBox& bbox, int padding, int mapWidth, int mapHeight);
